Split client
